refactor: move validation decorator to scores.utils
